+++ src/main/java/com/pluralsight/BakeHouse05/BaseRoom/Price.java
@@ -1,4 +1,4 @@
-package com.pluralsight.BakeHouse05;
+package com.pluralsight.BakeHouse05.BaseRoom;
 
 public interface Price {
     double calculatePrice();
