+++ src/main/java/com/pluralsight/BakeHouse05/Models/CakeSize.java
@@ -1,4 +1,4 @@
-package com.pluralsight.BakeHouse05;
+package com.pluralsight.BakeHouse05.Models;
 
 public enum CakeSize {
     SMALL("Small (6\")", 15.99),
