+++ src/main/java/com/pluralsight/BakeHouse05/Models/ToppingCategory.java
@@ -1,4 +1,4 @@
-package com.pluralsight.BakeHouse05;
+package com.pluralsight.BakeHouse05.Models;
 
 public enum ToppingCategory {
     REGULAR,
